add level drafts
